requirements.txt
anyio==4.4.0
asgiref==3.8.1
attrs==22.2.0
beautifulsoup4==4.11.2
certifi==2024.6.2
cffi==1.16.0
cfgv==3.4.0
charset-normalizer==3.3.2
cryptography==42.0.7
defusedxml==0.8.0rc2
distlib==0.3.8
Django==5.0.6
django-bootstrap5==22.2
django-colorfield==0.11.0
django-cors-headers==4.3.1
django-filter==24.2
django-templated-mail==1.1.1
django_debug_toolbar==3.8.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
drf-extra-fields==3.7.0
drf-yasg==1.21.7
drf_base64==2.0
Faker==12.0.1
filelock==3.14.0
filetype==1.2.0
flake8==5.0.4
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
identify==2.5.36
idna==3.7
inflection==0.5.1
iniconfig==2.0.0
mccabe==0.7.0
mixer==7.2.2
nodeenv==1.8.0
oauthlib==3.2.2
packaging==23.0
pillow==10.3.0
platformdirs==4.2.1
pluggy==1.0.0
pre-commit==3.7.0
psycopg2-binary==2.9.9
py==1.11.0
pycodestyle==2.9.1
pycparser==2.22
pyflakes==2.5.0
PyJWT==2.8.0
pypng==0.20220715.0
pytest==7.1.3
pytest-django==4.5.2
python-dateutil==2.8.2
python-dotenv==1.0.1
python-telegram-bot==21.3
python3-openid==3.2.0
pytz==2022.7
PyYAML==6.0.1
qrcode==7.4.2
requests==2.31.0
requests-oauthlib==2.0.0
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
six==1.16.0
sniffio==1.3.1
social-auth-app-django==5.4.1
social-auth-core==4.5.4
soupsieve==2.4.1
sqlparse==0.5.0
tomli==2.0.1
typing_extensions==4.12.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
virtualenv==20.26.1
watchdog==4.0.1
yapf==0.32.0
